+++ Albot/Code/Graphics/Model.cpp
@@ -28,7 +28,7 @@ Buffer* Model::get_index_buffer()
 
 uint32_t Model::get_index_total_count() const
 {
-	return m_triangle_indices.size() * 3;
+	return static_cast<uint32_t>(m_triangle_indices.size() * 3);
 }
 
 void Model::InitBuffer(DXRenderer* dxrenderer)
@@ -41,7 +41,7 @@ void Model::InitBuffer(DXRenderer* dxrenderer)
 		vb_buffer_desc.m_desc.m_usageType = Usage_Type::USAGE_IMMUTABLE;
 		vb_buffer_desc.m_desc.m_vertexStride = sizeof(VertexData);
 		vb_buffer_desc.m_rawData = &m_vertices[0];
-		vb_buffer_desc.m_size = sizeof(VertexData) * m_vertices.size();
+		vb_buffer_desc.m_size = static_cast<uint32_t>(sizeof(VertexData) * m_vertices.size());
 
 		m_vertex_buffer = DXResourceLoader::Create_Buffer(dxrenderer, vb_buffer_desc);
 	}
@@ -54,7 +54,7 @@ void Model::InitBuffer(DXRenderer* dxrenderer)
 		indx_buffer_desc.m_desc.m_usageType = Usage_Type::USAGE_IMMUTABLE;
 		indx_buffer_desc.m_desc.m_indexType = IndexType::INDEX_UINT32;
 		indx_buffer_desc.m_rawData = &m_triangle_indices[0];
-		indx_buffer_desc.m_size = sizeof(Triangle) * m_triangle_indices.size();
+		indx_buffer_desc.m_size = static_cast<uint32_t>(sizeof(Triangle) * m_triangle_indices.size());
 
 		m_index_buffer = DXResourceLoader::Create_Buffer(dxrenderer, indx_buffer_desc);
 	}
